=added clone function to age specific rate functions
the intention is to have any model be fit exactly once, and if there are parameters to vary,
to priors to be changed, then the model will be cloned.  this will produce a history, and allow
users to keep track of all the variants that they try when performing a fit.