Fix bug introduced in last commit
there are a few additional tweaks mixed in, but some how removing the initial conditions
in the last commit totally ruined the convergence of the mcmc.  this is particularly strange,
because the value is reseeded.  maybe adaptive metropolis guesses the proposal covariance
matrix from the current point?  i don't know why this makes such a difference.