Implement File Caching in DiseaseModelParameters
This will greatly speed up the performance of the website, since it will not need to generate complicated plots more than once.

This commit requires a migration, see docs/development.rst for directions.

TODO: the cache needs to be cleared when the plots change, to make sure the figures are updated